assignment-2/requirement.txt
flask
pymongo
dnspython
python-dotenv


